packages/api/requirements.txt
fastapi==0.115.*
uvicorn[standard]==0.34.*
celery[redis]==5.5.*
redis==5.2.*
supabase==2.*
openai==1.*
httpx==0.28.*
pymupdf==1.25.*
python-docx==1.1.*
trafilatura==2.*
python-multipart==0.0.*
pydantic-settings==2.*
